+++ R/summary.R
@@ -36,8 +36,8 @@
 #' diverted to \code{\link{hpd.summary}}.
 #' 
 #' \code{confint.emmGrid} is equivalent to \code{summary.emmGrid with 
-#' infer = c(TRUE, FALSE)}. When called with \code{joint = FALSE}, \code{test.emmGrid}
-#' is equivalent to \code{summary.emmGrid} with \code{infer = c(FALSE, TRUE)}. 
+#' infer = c(TRUE, FALSE)}. The function \code{test.emmGrid}, when called with 
+#' \code{joint = FALSE}, is equivalent to \code{summary.emmGrid} with \code{infer = c(FALSE, TRUE)}. 
 #' 
 #' With \code{joint = TRUE}, \code{test.emmGrid} calculates the Wald test of the
 #' hypothesis \code{linfct \%*\% bhat = null}, where \code{linfct} and
@@ -46,7 +46,8 @@
 #' non-estimable. It is permissible for the rows of \code{linfct} to be linearly
 #' dependent, as long as \code{null == 0}, in which case a reduced set of 
 #' contrasts is tested. Linear dependence and nonzero \code{null} cause an 
-#' error.
+#' error. The returned object has an aditional \code{"est.fcns"} attribute, which
+#' is a list of the linear functions associated with the joint test.
 #'
 #' @param object An object of class \code{"emmGrid"} (see \link{emmGrid-class})
 #' @param infer A vector of one or two logical values. The first determines
@@ -61,6 +62,9 @@
 #' @param by Character name(s) of variables to use for grouping into separate 
 #'   tables. This affects the family of tests considered in adjusted \emph{P}
 #'   values. 
+#' @param cross.adjust Character: \eqn{p}-value adjustment method to 
+#'   additionally apply \emph{across} 
+#'   the \code{by} groups. See the section on P-value adjustments for details.
 #' @param type Character: type of prediction desired. This only has an effect if
 #'   there is a known transformation or link function. \code{"response"} 
 #'   specifies that the inverse transformation be applied. \code{"mu"} (or 
@@ -217,6 +221,28 @@
 #'   \emph{Warning:} Non-estimable cases are \emph{included} in the family to which adjustments
 #'   are applied. You may wish to subset the object using the \code{[]} operator
 #'   to work around this problem.
+#'   
+#'   The \code{cross.adjust} argument is a way of specifying a multiplicity
+#'   adjustment across the \code{by} groups (otherwise by default, each group is
+#'   treated as a separate family in regards to multiplicity adjustments). It
+#'   applies only to \eqn{p} values. Valid options are one of the
+#'   \code{p.adjust.methods} or \code{"sidak"}. This argument is ignored unless
+#'   it is other than \code{"none"}, there is more than one \code{by} group, and
+#'   they are all the same size. Under those conditions, we first use
+#'   \code{adjust} to determine the within-group adjusted \eqn{p} values.
+#'   Imagine each group's adjusted \eqn{p} values arranged in side-by-side
+#'   columns, thus forming a matrix with the number of columns equal to the
+#'   number of \code{by} groups. Then we use the \code{cross.adjust} method to
+#'   further adjust the adjusted \eqn{p} values in each row of this matrix. Note
+#'   that an \emph{overall} Bonferroni (or Sidak) adjustment is obtainable by
+#'   specifying \emph{both} \code{adjust} and \code{cross.adjust} as
+#'   \code{"bonferroni"} (or \code{"sidak"}). However, less conservative (but
+#'   yet conservative) overall adjustments are available when it is possible to
+#'   use an \dQuote{exact} within-group method (e.g., \code{adjust = "tukey"}
+#'   for pairwise comparisons) and \code{cross.adjust} as a conservative
+#'   adjustment. [\code{cross.adjust} methods other than \code{"none"}, 
+#'   \code{"bonferroni"}, or \code{"sidak"} do not seem advisable, but other 
+#'   \code{p.adjust} methods are available if you can make sense of them.]
 #'
 #' @section Tests of significance, nonsuperiority, noninferiority, or equivalence:
 #'   When \code{delta = 0}, test statistics are the usual tests of significance.
@@ -332,10 +358,22 @@
 #' # Consider as some of many possible contrasts among the six cell means
 #' summary(con, infer = c(TRUE, TRUE), adjust = "scheffe", scheffe.rank = 5)
 #'
-summary.emmGrid <- function(object, infer, level, adjust, by, type, df, calc,
-                        null, delta, side, frequentist, 
-                        bias.adjust = get_emm_option("back.bias.adj"),
-                        sigma, ...) {
+#' # Show estimates to more digits
+#' print(test(con), digits = 7)
+#' 
+#' # --------------------------------------------------------------
+#' # Cross-adjusting P values
+#' prs <- pairs(warp.emm)   # pairwise comparisons of tension, by wool
+#' test(prs, adjust = "tukey", cross.adjust = "bonferroni")
+#' 
+#' # Same comparisons taken as one big family (more conservative)
+#' test(prs, adjust = "bonferroni", by = NULL)
+#' 
+summary.emmGrid <- function(object, infer, level, adjust, by, 
+                            cross.adjust = "none", type, df, calc,
+                            null, delta, side, frequentist, 
+                            bias.adjust = get_emm_option("back.bias.adj"),
+                            sigma, ...) {
     if(missing(sigma))
         sigma = object@misc$sigma
     if(missing(frequentist) && !is.null(object@misc$frequentist))
@@ -599,6 +637,27 @@ summary.emmGrid <- function(object, infer, level, adjust, by, type, df, calc,
             # we ignore everything about apv except the message
             mesg = c(mesg, apv$mesg)
         }
+        
+        # Handle cross-adjustments
+        if ( (length(by.rows) > 1) && 
+             (length(len <- sapply(by.rows, length)) > 1) &&
+             is.na(pmatch(cross.adjust, "none")) ) {
+            val = c("sidak", p.adjust.methods)
+            w = pmatch(tolower(cross.adjust), tolower(val))
+            if (!is.na(w)) {
+                cross.adjust = val[w]
+                mat = matrix(result$p.value, nrow = len)
+                apv = apply(mat, 1, function(p) {
+                    if (w > 1)   p.adjust(p, cross.adjust)
+                    else         1 - (1 - p)^ncol(mat)
+                })
+                result$p.value = as.numeric(t(apv))
+                mesg = c(mesg, paste("Cross-group P-value adjustment:", cross.adjust))
+            }
+            else
+                message("Invalid cross-adjustment method: '", cross.adjust, "'")
+        }
+        
         if (delta > 0)
             mesg = c(mesg, paste("Statistics are tests of", c("nonsuperiority","equivalence","noninferiority")[side+2],
                                  "with a threshold of", signif(delta, 5)))
@@ -721,9 +780,6 @@ predict.emmGrid <- function(object, type,
 #'   \code{emm_options(opt.digits = FALSE)}.
 #' @method as.data.frame emmGrid
 #' @export
-#' @examples
-#' # Show estimates to more digits
-#' print(test(con), digits = 7)
 as.data.frame.emmGrid = function(x, row.names = NULL, optional, check.names = TRUE, ...) {
     as.data.frame(summary(x, ...), row.names = row.names, check.names = check.names)
 }